added trainindg based on latent representations
