+++ experiments/__init__.py
@@ -1,2 +1,3 @@
 from .clustering_ml_raw import clustering_ml_raw
 from .clustering_dl_raw import clustering_dl_raw
+from .clustering_dl_latent import clustering_dl_latent
