+++ main.py
@@ -29,16 +29,17 @@ def main(args):
     elif args.experiment_name == "clustering_dl_raw":
         experiments.clustering_dl_raw(train_ds, test_ds, log_dir, args)
 
-    # elif args.experiment_name == "clustering_dl_latent":
-    #     experiments.clustering.cluster_raw_signals_ml(train_ds, test_ds, log_dir)
+    elif args.experiment_name == "clustering_dl_latent":
+        experiments.clustering_dl_latent(train_ds, test_ds, log_dir, args)
+
     # elif args.experiment_name == "clustering_dl_latent_with_attention":
     #     experiments.clustering.cluster_raw_signals_ml(train_ds, test_ds, log_dir)
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
-    parser.add_argument('--dataset-config-file', type=str, default="./config/biotac2.yaml")
-    parser.add_argument('--experiment-name', type=str, default="clustering_dl_raw")
+    parser.add_argument('--dataset-config-file', type=str, default="./config/touching.yaml")
+    parser.add_argument('--experiment-name', type=str, default="clustering_dl_latent")
     parser.add_argument('--epochs', type=int, default=1000)
     parser.add_argument('--batch-size', type=int, default=256)
     parser.add_argument('--dropout', type=float, default=0.2)
