+++ models/autoencoders/__init__.py
@@ -1,2 +1,3 @@
 from .rnn import RCNNAutoencoder, RCNNAutoencoderConfig
 from .sae import TimeSeriesAutoencoder, TimeSeriesAutoencoderConfig
+from . import ops
